feat: support `now alias` reading config.alias
Reads the last deployment and will automatically alias to the package.json's config.alias property
