+++ bin/now-alias
@@ -9,6 +9,7 @@ import login from '../lib/login';
 import * as cfg from '../lib/cfg';
 import { error } from '../lib/error';
 import toHost from '../lib/to-host';
+import readMetaData from '../lib/read-metadata';
 
 const argv = minimist(process.argv.slice(2), {
   string: ['config', 'token'],
@@ -82,7 +83,7 @@ const exit = (code) => {
   setTimeout(() => process.exit(code || 0), 100);
 };
 
-if (argv.help || !subcommand) {
+if (argv.help) {
   help();
   exit(0);
 } else {
@@ -202,16 +203,20 @@ async function run (token) {
       break;
 
     default:
-      if (2 === argv._.length) {
-        await alias.set(String(argv._[0]), String(argv._[1]));
-      } else if (argv._.length >= 3) {
-        error('Invalid number of arguments');
-        help();
-        exit(1);
+      if (0 === argv._.length) {
+        await realias(alias);
       } else {
-        error('Please specify a valid subcommand: ls | set | rm');
-        help();
-        exit(1);
+        if (2 === argv._.length) {
+          await alias.set(String(argv._[0]), String(argv._[1]));
+        } else if (argv._.length >= 3) {
+          error('Invalid number of arguments');
+          help();
+          exit(1);
+        } else {
+          error('Please specify a valid subcommand: ls | set | rm');
+          help();
+          exit(1);
+        }
       }
   }
 
@@ -273,4 +278,23 @@ function findAlias (alias, list) {
   return _alias;
 }
 
+async function realias (alias) {
+  const path = process.cwd();
+  const { pkg, name, description } = await readMetaData(path, {
+    deploymentType: 'npm', // FIXME: hard coding settings…
+    quiet: true, // `quiet`
+  });
 
+  const pkgConfig = pkg ? pkg.config || {} : {};
+  const target = pkgConfig.alias;
+
+  if (!target) {
+    help();
+    return exit(0);
+  }
+
+  // now try to find the last deployment
+  const source = await alias.last(name);
+
+  await alias.set(source.url, target);
+}
