+++ lib/index.js
@@ -4,6 +4,7 @@ import {
   npm as getNpmFiles,
   docker as getDockerFiles
 } from './get-files';
+import readMetaData from './read-metadata';
 import ua from './ua';
 import hash from './hash';
 import retry from 'async-retry';
@@ -45,98 +46,20 @@ export default class Now extends EventEmitter {
   }) {
     this._path = path;
 
-    let pkg = {};
-    let name, description, files;
+    let files;
 
+    const { pkg, name, description } = await readMetaData(path, {
+      deploymentType,
+      quiet,
+    });
+
+    if (this._debug) console.time('> [debug] Getting files');
     if ('npm' === deploymentType) {
-      try {
-        pkg = await readFile(resolvePath(path, 'package.json'));
-        pkg = JSON.parse(pkg);
-      } catch (err) {
-        const e = Error(`Failed to read JSON in "${path}/package.json"`);
-        e.userError = true;
-        throw e;
-      }
-
-      if (!pkg.scripts || (!pkg.scripts.start && !pkg.scripts['now-start'])) {
-        const e = Error('Missing `start` (or `now-start`) script in `package.json`. ' +
-          'See: https://docs.npmjs.com/cli/start.');
-        e.userError = true;
-        throw e;
-      }
-
-      if (null == pkg.name || 'string' !== typeof pkg.name) {
-        name = basename(path);
-        if (!quiet) console.log(`> No \`name\` in \`package.json\`, using ${chalk.bold(name)}`);
-      } else {
-        name = pkg.name;
-      }
-
-      description = pkg.description;
-
-      if (this._debug) console.time('> [debug] Getting files');
-      files = await getNpmFiles(path, pkg, { debug: this._debug });
-      if (this._debug) console.timeEnd('> [debug] Getting files');
-    } else if ('docker' === deploymentType) {
-      let docker;
-      try {
-        const dockerfile = await readFile(resolvePath(path, 'Dockerfile'), 'utf8');
-        docker = parseDockerfile(dockerfile, { includeComments: true });
-      } catch (err) {
-        const e = Error(`Failed to parse "${path}/Dockerfile"`);
-        e.userError = true;
-        throw e;
-      }
-
-      if (!docker.length) {
-        const e = Error('No commands found in `Dockerfile`');
-        e.userError = true;
-        throw e;
-      }
-
-      if (!docker.some((cmd) => 'CMD' === cmd.name)) {
-        const e = Error('No `CMD` found in `Dockerfile`. ' +
-          'See: https://docs.docker.com/engine/reference/builder/#/cmd');
-        e.userError = true;
-        throw e;
-      }
-
-      if (!docker.some((cmd) => 'EXPOSE' === cmd.name)) {
-        const e = Error('No `EXPOSE` found in `Dockerfile`. A port must be supplied. ' +
-          'See: https://docs.docker.com/engine/reference/builder/#/expose');
-        e.userError = true;
-        throw e;
-      }
-
-      const labels = {};
-      docker
-      .filter(cmd => 'LABEL' === cmd.name)
-      .forEach(({ args }) => {
-        for (let key in args) {
-          // unescape and convert into string
-          try {
-            labels[key] = JSON.parse(args[key]);
-          } catch (err) {
-            const e = Error(`Error parsing value for LABEL ${key} in \`Dockerfile\``);
-            e.userError = true;
-            throw e;
-          }
-        }
-      });
-
-      if (null == labels.name) {
-        name = basename(path);
-        if (!quiet) console.log(`> No \`name\` LABEL in \`Dockerfile\`, using ${chalk.bold(name)}`);
-      } else {
-        name = labels.name;
-      }
-
-      description = labels.description;
-
-      if (this._debug) console.time('> [debug] Getting files');
+      files = await getNpmFiles(path, { debug: this._debug });
+    } else {
       files = await getDockerFiles(path, { debug: this._debug });
-      if (this._debug) console.timeEnd('> [debug] Getting files');
     }
+    if (this._debug) console.timeEnd('> [debug] Getting files');
 
     const nowProperties = pkg ? pkg.now || {} : {};
 
@@ -352,6 +275,23 @@ export default class Now extends EventEmitter {
     }, { retries: 3, minTimeout: 2500, onRetry: this._onRetry });
 
     return deployments;
+  }
+
+  async last (app) {
+    // TODO app is required
+    const deployments = await this.list(app);
+
+    const last = deployments.sort((a, b) => {
+      return b.created - a.created;
+    }).shift();
+
+    if (!last) {
+      const e = Error(`No deployments found for "${app}"`);
+      e.userError = true;
+      throw e;
+    }
+
+    return last;
   }
 
   async listAliases (deploymentId) {
